Added loaders for different datasets, replaced the neural network with a deeper CNN, and trained a baseline adversarially trained model.
Replaced most of the contents of the Adversarial Sample Detection with Clustering notebook with utility functions defined in utils.py,
and implemented the procedure of detecting adversarial samples as a class which is imported in the notebook. Notebook is where I have
been testing the Clustering Adversarial Classifier and comparing it with the baseline. Also slightly modifier the procedure of the
clustering adversarial detector so that it provides predictions on samples flagged as suspicious, making it more of a traditional classifer.