Added LayerwiseClustering class and View_Sample_At_Each_Layer notebook to test the class.
LayerwiseClustering class saves the input to convolutional and linear layers so that they
can be clustered or visualized later. Currently this is used to view how an adversarial sample
drifts with respect to its original class as it gets deeper into the attacked model. This is
shown in the View_Sample_At_Each_Layer notebook.